feat: add more scripts
